Delete SNAIL_RENDERER_XXX macros in snail header files
...to allows .cpp files in module other than snail which include snail
headers to be compiled without SNAIL_RENDERER_XXX macros.
