+++ CMakeLists.txt
@@ -278,7 +278,6 @@ if(ANDROID)
   set(EXTERNAL_LIB_LOCATION ${CMAKE_SOURCE_DIR}/android/external)
 
   SET(SNAIL_HEADLESS OFF CACHE BOOL "Snail SDL")
-  add_definitions(-DSNAIL_RENDERER_SDL)
 
   add_library( SDL2 SHARED IMPORTED )
   add_library( SDL2_image SHARED IMPORTED )
@@ -329,11 +328,9 @@ else()
   if((WITH_TESTS STREQUAL "TESTS") OR (WITH_TESTS STREQUAL "BENCH"))
     unset(SNAIL_HEADLESS CACHE)
     SET(SNAIL_HEADLESS ON CACHE BOOL "Snail headless")
-    add_definitions(-DSNAIL_RENDERER_HEADLESS)
   elseif(WITH_TESTS STREQUAL "OFF")
     unset(SNAIL_HEADLESS CACHE)
     SET(SNAIL_HEADLESS OFF CACHE BOOL "Snail SDL")
-    add_definitions(-DSNAIL_RENDERER_SDL)
   else()
     # Use the previous flags.
   endif()
