+++ src/snail/detail/sdl_headless.cpp
@@ -5,8 +5,6 @@ namespace snail
 namespace detail
 {
 
-
-
 void enforce_sdl(int)
 {
 }
@@ -25,6 +23,34 @@ void enforce_image(int)
 
 void enforce_mixer(int)
 {
+}
+
+
+
+void* enforce_sdl_internal(void* result)
+{
+    return result;
+}
+
+
+
+void* enforce_ttf_internal(void* result)
+{
+    return result;
+}
+
+
+
+void* enforce_img_internal(void* result)
+{
+    return result;
+}
+
+
+
+void* enforce_mixer_internal(void* result)
+{
+    return result;
 }
 
 
@@ -74,7 +100,6 @@ SDLMixer::SDLMixer()
 SDLMixer::~SDLMixer()
 {
 }
-
 
 } // namespace detail
 } // namespace snail
